requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.0
Flask-RESTful==0.3.10
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.172.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.2
googlemaps==4.10.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huepy==1.2.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.7.1
markdown2==2.5.3
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.4
oauthlib==3.2.0
openai==1.86.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
PyMuPDF==1.25.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
PyYAML==5.4.1
reportlab==4.3.1
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
systemd-python==234
texttable==1.6.4
tornado==6.4.2
tqdm==4.67.1
Twisted==22.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==1.26.5
virtualenv==20.29.3
vite==1.5.2
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
zipp==1.0.0
zope.interface==5.4.0
